requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
pymongoose==0.1
requests==2.31.0
python-dotenv==0.19.0
flask-cors==3.0.10
passlib==1.7.4
PyJWT==2.3.0